FastAPI機能をモジュール切り分け
