+++ virtual_pertner_back/fastAPI/.gitignore
@@ -1,2 +1,3 @@
 # Env files
 .env
+.wav
