+++ virtual_pertner_back/fastAPI/app/__init__.py
@@ -1,0 +1,1 @@
+# FastAPI/app/__init__.py
