+++ virtual_pertner_back/fastAPI/app/routes/__init__.py
@@ -1,0 +1,5 @@
+# FastAPI/app/routes/__init__.py
+
+from .chat import router as chat_router
+
+__all__ = ["chat_router"]
