+++ virtual_pertner_back/fastAPI/app/services/__init__.py
@@ -1,0 +1,1 @@
+# FastAPI/app/services/__init__.py
